src: cpu: conv: gemm_x8s8s32x, jit_brgemm_1x1, jit_brgemm: fix warnings

+++ src/cpu/gemm_x8s8s32x_convolution.cpp
@@ -357,7 +357,7 @@ status_t gemm_x8s8s32x_convolution_bwd_data_t::execute_backward_data_thr(
             = types::data_type_size(diff_src_md.data_type());
 
     const int scale_idx_mult = pd()->attr()->scales_.get(DNNL_ARG_WEIGHTS).mask_
-            == (1 << pd()->with_groups());
+            == (1 << static_cast<int>(pd()->with_groups()));
     DEFINE_ARG_SCALES_BUFFER(src_scales, DNNL_ARG_SRC);
     DEFINE_ARG_SCALES_BUFFER(wei_scales, DNNL_ARG_WEIGHTS);
     DEFINE_ARG_SCALES_BUFFER(dst_scales, DNNL_ARG_DST);
